Move classes into sub-package
